feat: 네비게이션 가드
